mission_control_backend/requirements.txt
actionlib==1.14.0
angles==1.9.13
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.10.4
backports.zoneinfo==0.2.1
base-local-planner==1.17.3
bondpy==1.8.6
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
dnspython==2.6.1
dynamic-reconfigure==1.7.3
exceptiongroup==1.3.0
fastapi==0.104.1
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
image-geometry==1.16.2
interactive-markers==1.12.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
laser-geometry==1.6.7
message-filters==1.17.0
paho-mqtt==2.1.0
pydantic==2.5.1
pydantic-core==2.14.3
pymongo==4.10.1
pyproj==3.5.0
python-dateutil==2.9.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-qt-binding==0.4.4
pytz==2025.2
PyYAML==6.0.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.32.4
resource-retriever==1.12.7
rosbag==1.17.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.17.0
roslaunch==1.17.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.17.0
rosmake==1.15.8
rosmaster==1.17.0
rosmsg==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospy==1.17.0
rosservice==1.17.0
rostest==1.17.0
rostopic==1.17.0
rosunit==1.15.8
roswtf==1.17.0
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.12
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.17
rqt-launch==0.4.9
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.4
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rviz==1.14.25
sensor-msgs==1.13.1
six==1.17.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
sniffio==1.3.1
starlette==0.27.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
topic-tools==1.17.0
typing-extensions==4.13.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.24.0
websockets==12.0
xacro==1.14.19
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
